requirements.txt
blinker==1.9.0
catboost==1.2.7
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
geographiclib==2.0
geopy==2.4.1
graphviz==0.20.3
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightgbm==4.5.0
LinearRegression==0.0.1
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.1.3
xgboost==2.1.3
zipp==3.21.0
python-dotenv==1.0.0
google-api-python-client==2.91.0